* Cleanup, use longs rather than ints where appropriate * Re-enabled SoundTouch code, if Mixxx crashes on you and you figure out the bug in lines 174-192 (enginebufferscalest.cpp), I will give you $5 * Bonus: I will give you an extra $5 if I'm wrong and the bug isn't where I said it is * Added ASSERT to EngineBufferScaleST to help track down problems

+++ mixxx/src/enginebufferscale.h
@@ -45,7 +45,7 @@ public:
     /** Called from EngineBuffer when seeking, to ensure the buffers are flushed */
     virtual void clear() = 0;
     /** Scale buffer */
-    virtual CSAMPLE *scale(double playpos, int buf_size, float *pBase=0, int iBaseLength=0) = 0;
+    virtual CSAMPLE *scale(double playpos, unsigned long buf_size, float *pBase=0, unsigned long iBaseLength=0) = 0;
     
 protected:
     /** Pointer to ReaderExtractWave object */
